requirements.txt
alembic==1.8.1
bcrypt==4.0.1
click==8.1.3
coverage==6.5.0
dataclasses-json==0.5.7
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
itsdangerous==2.0.1
mypy==0.982
pika==1.3.1
psycopg2-binary==2.9.5
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dotenv==0.21.0
SQLAlchemy==1.4.42
WTForms==3.0.1
